test: move helm tests into their own file, and skip them on short tests
On windows, running Helm can take ~0.5 seconds,
which starts to blow up test times. Hoping this
will make the tests less flakey
