+++ internal/tiltfile/tiltfile_docker_compose_test.go
@@ -1,4 +1,4 @@
-//+build !skipdockercomposetests
+//+build !skiplargetiltfiletests
 
 package tiltfile
